refactor: fix and clean up independent URI resolvers
