+++ mycore-base/src/main/java/org/mycore/common/xsl/uriresolver/MCRFileResolver.java
@@ -1,0 +1,50 @@
+/*
+ * This file is part of ***  M y C o R e  ***
+ * See https://www.mycore.de/ for details.
+ *
+ * MyCoRe is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MyCoRe is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.mycore.common.xsl.uriresolver;
+
+import java.io.IOException;
+import java.net.URI;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+import javax.xml.transform.Source;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.URIResolver;
+import javax.xml.transform.stream.StreamSource;
+
+public class MCRFileResolver implements URIResolver {
+
+    @Override
+    public Source resolve(String href, String base) throws TransformerException {
+        URI hrefURI = MCRURIResolver.resolveURI(href, base);
+        if (!hrefURI.getScheme().equals("file")) {
+            throw new TransformerException("Unsupport file uri scheme: " + hrefURI.getScheme());
+        }
+        Path path = Paths.get(hrefURI);
+        StreamSource source;
+        try {
+            source = new StreamSource(Files.newInputStream(path), hrefURI.toASCIIString());
+            return source;
+        } catch (IOException e) {
+            throw new TransformerException(e);
+        }
+    }
+
+}
