+++ mycore-base/src/main/java/org/mycore/common/xsl/uriresolver/MCRPropertiesResolver.java
@@ -96,4 +96,5 @@ public class MCRPropertiesResolver implements URIResolver {
 
         return new JDOMSource(entryElement);
     }
+
 }
