+++ mycore-base/src/main/java/org/mycore/common/xsl/uriresolver/MCRURIResolverFilter.java
@@ -218,4 +218,5 @@ public class MCRURIResolverFilter implements Filter {
 
         }
     }
+
 }
