+++ mycore-base/src/main/java/org/mycore/common/xsl/uriresolver/MCRVersionInfoResolver.java
@@ -1,0 +1,84 @@
+/*
+ * This file is part of ***  M y C o R e  ***
+ * See https://www.mycore.de/ for details.
+ *
+ * MyCoRe is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MyCoRe is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.mycore.common.xsl.uriresolver;
+
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.List;
+
+import javax.xml.transform.Source;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.URIResolver;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.jdom2.Element;
+import org.jdom2.transform.JDOMSource;
+import org.mycore.common.xml.MCRXMLFunctions;
+import org.mycore.datamodel.common.MCRAbstractMetadataVersion;
+import org.mycore.datamodel.common.MCRXMLMetadataManager;
+import org.mycore.datamodel.metadata.MCRObjectID;
+
+public class MCRVersionInfoResolver implements URIResolver {
+
+    private static final Logger LOGGER = LogManager.getLogger();
+
+    @Override
+    public Source resolve(String href, String base) throws TransformerException {
+        String id = href.substring(href.indexOf(':') + 1);
+        LOGGER.debug("Reading version info of MCRObject with ID {}", id);
+        MCRObjectID mcrId = MCRObjectID.getInstance(id);
+        MCRXMLMetadataManager metadataManager = MCRXMLMetadataManager.obtainInstance();
+        try {
+            List<? extends MCRAbstractMetadataVersion<?>> versions = metadataManager.listRevisions(mcrId);
+            if (versions != null && !versions.isEmpty()) {
+                return getSource(versions);
+            } else {
+                return getSource(Instant.ofEpochMilli(metadataManager.getLastModified(mcrId))
+                    .truncatedTo(ChronoUnit.MILLIS));
+            }
+        } catch (Exception e) {
+            throw new TransformerException(e);
+        }
+    }
+
+    @SuppressWarnings("PMD.AvoidDuplicateLiterals")
+    private Source getSource(Instant lastModified) {
+        Element e = new Element("versions");
+        Element v = new Element("version");
+        e.addContent(v);
+        v.setAttribute("date", lastModified.toString());
+        return new JDOMSource(e);
+    }
+
+    @SuppressWarnings("PMD.AvoidDuplicateLiterals")
+    private Source getSource(List<? extends MCRAbstractMetadataVersion<?>> versions) {
+        Element e = new Element("versions");
+        for (MCRAbstractMetadataVersion<?> version : versions) {
+            Element v = new Element("version");
+            v.setAttribute("user", version.getUser());
+            v.setAttribute("date", MCRXMLFunctions.getISODate(version.getDate(), null));
+            v.setAttribute("r", version.getRevision());
+            v.setAttribute("action", Character.toString(version.getType()));
+            e.addContent(v);
+        }
+        return new JDOMSource(e);
+    }
+
+}
